backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
arxiv==2.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-genai==0.3.0
pydantic==2.6.4
